change duration to long
git-svn-id: https://svn.win.tue.nl/repos/prom/Packages/InductiveVisualMiner/Trunk@43795 a68f48d4-8d00-424f-b812-3f22a534a02a

+++ src/org/processmining/plugins/inductiveVisualMiner/helperClasses/ResourceTimeUtils.java
@@ -46,11 +46,11 @@ public class ResourceTimeUtils {
 		return String.format("%.0f", d) + "d";
 	}
 
-	public static final String getDurationPadded(double ms) {
-		if (ms / 1000 < 1) {
+	public static final String getDurationPadded(long ms) {
+		if (ms / 1000.0 < 1) {
 			return String.format("%.2f", ms) + " ms  ";
 		}
-		double s = ms / 1000;
+		double s = ms / 1000.0;
 		if (s / 60 < 1) {
 			return String.format("%.2f", s) + " s   ";
 		}
